get ready to load ipfs URLs
There is the issue that ipfs URLs are currently case-sensitive with
base58 hashes, but this has been recognized as a problem:
https://github.com/ipfs/ipfs/issues/337
QUrl also converts addresses to lowercase.  So we can't currently
load base58 hashes into QUrl; consequently switched load(QUrl)
to load(QString) for now.  Maybe it can be reverted when they
switch to base32 hashes.