fix the back button
QUrl.isRelative() returns true for IPFS hashes.
We store complete resolved URLs on the m_history stack, and
it's less confusing if that's always the same as what is shown
in the URL field.  Always push the URL even if we don't load
successfully, since on_actionGo_back_triggered() will always
pop it off before going back to the url before.