nb: bring in thumbnail view and start introducing Page Series
A Page Series will simply be a series of pages linked together via IPLD
into a linked list.  It can be used to make presentations for example:
one file per slide, but be able to rearrange the order at any time.

Anyway I need a way to graphically manipulate the series, and this
thumbnail view/scene stuff I wrote for Taborca (to rearrange pages in
a PDF) might be a good start.