nb: save Page Series with IPLD
That is, create a DAG node with JSON containing something like
{"series":["Qmcid1","Qmcid2"]}
for a series of two documents.  Not sure if this is a good idea...
so far nothing knows what to do with it: nb can't read it back because
it uses KIO and that uses the files/read API, but it's not a file.
So maybe nb will have to do a pre-check of anything it tries to
load to see whether it's something special, or to show raw JSON
in the JSON viewer.