nb: add settings for saving resources, and settings dialog
Settings to control whether to save resources locally at all,
and if so, a suffix for a subdir to save them into.  E.g. if you
are saving document.md by default it will create a directory
document_resources under the directory where you're saving the
document, and put the images in there.

This doesn't work for saving to IPFS yet.