nb: preload CSS for HTML only
CSS styling of HTML is tightly coupled with the HTML parsing itself
in Qt (QTextDocument etc.)  For Markdown, CSS doesn't work at all yet,
except that I'm working on a patch to try to fix that... but it's not
completely straightforward.  So in the meantime, at least giving the
stylesheet before loading HTML makes it work for that.