show * in titlebar if modified; prompt to save; make file-watching more reliable
Prompt to save the file if attempting to quit or reload when it's modified.
Update the QFileSystemWatcher at appropriate intervals; it seems that
saving a file probably involves saving a new one and removing the old one,
which disrupts the watcher, so we need to update it afterwards.  But
don't prompt the user to reload the file if it changed as a result of
being saved, either.