Add feature to automatically move "done" tasks down
Settings permit moving the checked task to another list under a heading
labeled with "done" or some other synonym; or failing that, moving it
to the bottom of the same list it's already in.  This fits with my
intuitive way of handling todo lists before I discovered todo.txt
format or had the idea of using markdown for it.

So far this is done only when purposely triggering the toggle-checkbox
action, not when clicking on the checkbox itself.  I didn't figure
out yet how to be notified when that happens.

When the cursor is merely moved to a checked list item, that causes the
toggle-checkbox action to be programmatically toggled; so now we use the
triggered signal instead of toggled, to avoid automatic reactions in that case.

Control-Z to undo works, but takes several steps to put the task item
back where it came from.  Otherwise there's no attempt yet to
automatically move an unchecked task back where it came from, but it
should be possible; BlockUserData holds the previous position to
facilitate that.