Added the template feature
A journal entry can be based on a template in ~/journal/.templates/journal.md.
The purpose is to remind me to record some data points every day.
Data points can be put into tables or bullet points; perhaps later
I'll come up with a way to query across a series of .md files as if
they were rows in a database table.

Anyway, templates seem like they should be useful in general; so
now there is a -t option to specify the name of the template to load,
which will try to find a file in the .templates directory in the same
directory where the document is, or else ~/.config/NetteBook/templates
or /etc/xdg/Nettebook/templates.

To be consistent with that, settings are now moved to
~/.config/NetteBook/NetteBook.conf on Linux.  It remains to be seen
whether this needs attention on macOS.  I prefer to have all
local configuration in the same place (QTBUG-82888).