Merge a new list item into an existing list when possible
When you turn a plain-text paragraph into a list item by choosing a
list-item style, or by hitting the checked-state toolbar button or
hitting control-K, it will first try to append this item to the adjacent
list above, if possible; if not, then try to prepend this item to the
adjacent list below, if possible; if not, only then create a new list.

Before, on_styleCB_activated() was creating a new list unconditionally
whenever the chosen style is a list item.  That wasn't so evident when
writing markdown (and that might even be a bug in QTextMarkdownWriter),
but quite evident in HTML, and also interfering with an attempt to
automatically move checked tasks to the bottom.

Also change the square bullet character in the style combobox to
U+25AA BLACK SMALL SQUARE because the open square kept grabbing my
attention when I actually want to make it an unchecked task instead.