Beginnings of kanban mode: view and model for easy DnD editing
If an MD file has headings # To Do # Doing # Done (for example) with
task lists under each of those, you can view it in columns instead of
just as a document, and use DnD to move tasks between columns.
However, you are actually still editing the QTextDocument when
you are doing that.  The kanban view is currently a separate window
so you can watch how these actions edit the document; but probably
it should replace the main view.

You can have any number of columns.  Only the Done column is special:
it's where checked tasks should be moved, if that setting is turned on.

kanban.md is an example for this project, tracking some upcoming
tasks (not all of them are about developing kanban mode, either).