Allow DnD only between items, not onto them; fix DnD bugs
Attempting automatic insertion in all cases brings various Qt bugs to attention:
inserting at the end of a heading makes the next line a heading
you can't insert a block with its text and its format at the same time,
so we have to prepend or append a newline in cases where it's OK to
inherit the block format, and then modify it afterwards

Deletion of the existing row after a move had an off-by-one error
in the case that we move it to a position above where it was (naturally).

Anyway the workarounds seem to work now, so far.