Handle drop events on main window: single file only, in read-only mode
Just open the file URL that is dragged from the file manager (finder)
and dropped onto the window.  This is a good alternative especially on
macOS, since using command-line options is harder with an app bundle.
But it looks like I should convert this to a proper SDI application: if
you drop multiple files, they ought to open in multiple windows.  That's
not done yet; so for now it only takes QMimeData::urls().first().
Also, QTextBrowser in edit mode reacts by inserting the text taken from
the drop event; so opening files is best done in read-only mode.
This needs work.