add a pie menu; prepare to link selections
The pie menu is not quite identical to the qtdeclarative example:
it pops up instantly with a right click and slowly on touchscreen.
Probably worth upstreaming that change anyway.

There's only one instance, but we can find the TextEdit on which
it was opened and the selected range of text. Now to figure out
how to make links between notes...