Add ZkModel::getLinkedIndices()
This needs a provider function a la TableView.columnWidthProvider so
that ZkModel can get the QQuickTextDocument from the TextEdit that has
already been instantiated.  (Alternatively I suppose ZkModel could
create the QTDs ahead of time, but why? That's not its primary concern
so far.) (I seem to remember that there's supposed to be a better way to
write provider functions though, not using QJSValue.)

Whew that was a bit of work, but now each "card" can find out what other
"cards" it links to. Maybe later it should return a gadget instead of
just an index, which would provide the character ranges, and then the
geometry of that could maybe be gotten somehow... but that might require
some digging around and adding new API in Qt Quick.  I think there might
be a bug asking for something like that too.