Use FontFixedPitch property in addition to mono font
See also https://bugreports.qt.io/browse/QTBUG-103484 QTBUG-54623
and related bugs.

If we do not set m_monoFont, QTextCharFormat::setFontFixedPitch() is not
enough to make QTextEdit actually use a mono font; but if we do not call
setFontFixedPitch(), QTextMarkdownWriter::writeBlock() may not detect
that a mono font is in use, because it may be that the font returned
from QFontDatabase::systemFont(QFontDatabase::FixedFont) is a
proportional one. That may be a bug in Qt; but the fallback to fix
QTBUG-54623 is https://codereview.qt-project.org/c/qt/qtbase/+/260945

Amends df5ebff2da347a7d84a464c6cba562f3c8f99d43