Fix list item with link, by hand
I did this in zk: positioned text cursor at the end of the list item
and pressed Delete to pull the next paragraph up into the list item.
This should stay the same now after some recent Qt fixes.
But the link could've been an autolink, and Qt is currently not
detecting that the link text and the URL are the same and converting
it to an autolink.  (Not sure if it should though either)