Convert autolink by hand
I did that in vi, then ran zk again and let it autosave. Hopefully
it stays that way now: Qt keeps track of which links are autolinks.
But that was supposed to be fixed a while ago (QTBUG-94713) so I'm
not sure how the link was not written that way to begin with. Maybe
wrapping the list item messed it up somehow.