Position new files randomly; don't reposition otherwise
This was already supposed to work; but YamlDocument::parsed() is not emitted
if there's no FrontMatter, so setRandomPosition() wasn't called. So we add
needsPosition() to directly trigger setRandomPosition().

If the document isn't loaded, there is definitely no yaml yet; so don't
emit either signal until QQuickTextDocument::status() is Loaded.
(It's good that we have this signal since
https://codereview.qt-project.org/c/qt/qtdeclarative/+/536960 )

Since we depend on QQuickTextDocument anyway, there's actually no use
for the YamlDocument.source property. Remove it.